\name{jmosaics_example_data}
\docType{data}
\alias{jmosaics_example_data}
\title{
ChIP-seq data of Histone H3K4me1 and H3K27me3.}
\description{
This is an example chr10 ChIP-seq data of Histone H3K4me1 and H3K27me3.}
\usage{
data("jmosaics_example_data")
}
\format{
\code{jmosaics_example_data} names of included datasets.

}
\examples{

data(jmosaics_example_data)
jmosaics_example_data
}
\keyword{datasets}
