\name{origin_bin}
\docType{data}
\alias{origin_bin}
\title{
ChIP-seq data of Histone H3K4me1 and H3K27me3.}
\description{
This is an example chr10 ChIP-seq data of Histone H3K4me1 and H3K27me3.}
\usage{
data("jmosaics_example_data")
}
\format{
\code{origin_bin} a list of bin-level data sets.

}
\examples{

data(jmosaics_example_data)
str(origin_bin)
}
\keyword{datasets}
