\name{readBinsMultiple}
\alias{readBinsMultiple}

\title{
Match coordinates for multiple data sets
}
\description{
Match coordinates for multiple data sets}
\usage{
readBinsMultiple(dataset)
}
\arguments{
  \item{dataset}{
 A list of bin-level data sets.
}
}
\details{
Bin-level data can be generated by readBins from mosaics package for each data set.
This function is used to match the coordinates of multiple data sets to get the shared regions.

}
\value{
List of Bin-level data sets.
}

\author{
Xin Zeng
}
\examples{
data("jmosaics_example_data")
bin<- readBinsMultiple(origin_bin)

}

\keyword{ models }
\keyword{ methods }
